feat: price clicks in fractions of a token, x2 only from 70%
The ladder is now x1.25 from 25% of the map, x1.5 from 50% and x2 from
70%. A whole-token cost could not have a rung between x1 and x2, so the
toll, the limiter and ClickBudget.cost take a double. The proto moves
cost and next_cost to new field numbers: an old client reads zero and
says nothing about price instead of failing to decode.

+++ apps/backend/internal/planet/internal/clicks/usecases/click/throttle_click/usecase.go
@@ -23,7 +23,7 @@ import (
 // reading comes back either way: a refused caller is the one that most needs to
 // know when the next token lands.
 type Limiter interface {
-	TakeN(key string, n int) (bool, cpratelimit.State)
+	TakeN(key string, n float64) (bool, cpratelimit.State)
 }
 
 // Pricer says how many tokens a click for a country costs.
